removed & inlined custom View type structs for buttons
XCode doesn't like having custom structs like that, that inherit View, because it's not showing the bounds on each icon on the Preview Canvas anymore (which makes it very tough to debug why something doesn't look the way I want it to). I'm "inlining" those here again, because having the Preview Canvas actually show the rect of each item in the UI is too useful for me.