brought back TextFormattingButton
Was supposed to add this in the previews commit, but was too excited to change and fix this padding and spacing stuff which I saw after checking Canvas Preview, so I did that then and doing this now. Finally! It's some decent refactor and I can still see eveything correctly in Canvas Preview!